+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @relayroom/cli
 
+## 0.7.0
+
 ## 0.6.2
 
 ### Patch Changes